cidrlist: Rewrite file every time we call RecordIP
We call RecordIP when there is an IP that should be permitted.
Previously we would not update the file if the IP was already listed. We
do now. We do this by writing out to a temporary file and then moving it
into place.

We rewrite always for two reasons:
  1. Track last time an IP was seen (so old can be pruned)
  2. This causes iptables-manage, when running as a daemon, to sync the
     rules (since the file changed).